announce삭제시 announce에 신청한 application 모두 삭제 - 슬기

+++ src/main/java/kr/or/announce/service/AnnounceService.java
@@ -32,7 +32,14 @@ public class AnnounceService {
 	}
 
 	public int deleteAnnounce(int announceNo) {
-		return dao.deleteAnnounce(announceNo);
+		int appResult = dao.deleteApplication(announceNo);
+		if(appResult > 0) {
+			System.out.println("deleteAnnounce Dao 여기 성공");
+			return dao.deleteAnnounce(announceNo);
+		} else {
+			return 0;
+		}
+		
 	}
 
 	/*
